# packages/credential-showcase-openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Credential Showcase API
  version: 0.1.1
servers:
  - url: https://virtserver.swaggerhub.com/SphereonInt/showcase/0.1.0/v1
    description: SwaggerHub API Auto Mocking
  - url: https://api.example.com/v1
...
security:
  - OAuth2: []
  - OIDC: []
paths:
  /workflows/issuances:
    get:
      tags:
        - Issuance Flow
      summary: List all issuance flows
      operationId: listIssuanceFlows
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceFlowsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Issuance Flow
      summary: Create a new issuance flow
      operationId: createIssuanceFlow
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssuanceFlowRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceFlowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/issuances/{issuanceFlowId}:
    get:
      tags: [ "Issuance Flow" ]
      summary: Retrieve a single issuance flow
      operationId: getIssuanceFlow
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceFlowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Issuance Flow" ]
      summary: Update an issuance flow (overwrite)
      operationId: updateIssuanceFlow
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssuanceFlowRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceFlowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Issuance Flow" ]
      summary: Delete an issuance flow
      operationId: deleteIssuanceFlow
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/issuances/{issuanceFlowId}/steps:
    get:
      tags:
        - "Issuance Flow"
      summary: List steps in an issuance flow
      operationId: listIssuanceFlowSteps
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [ "Issuance Flow" ]
      summary: Add a step to an issuance flow
      operationId: createIssuanceFlowStep
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/issuances/{issuanceFlowId}/steps/{stepId}:
    get:
      tags: [ "Issuance Flow" ]
      summary: Retrieve a specific step in an issuance flow
      operationId: getIssuanceFlowStep
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Issuance Flow" ]
      summary: Overwrite a step in an issuance flow
      operationId: updateIssuanceFlowStep
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Step'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Issuance Flow" ]
      summary: Delete a step in an issuance flow
      operationId: deleteIssuanceFlowStep
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/issuances/{issuanceFlowId}/steps/{stepId}/actions:
    get:
      tags: [ "Issuance Flow" ]
      summary: List actions in a step
      operationId: listStepActions
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepActionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [ "Issuance Flow" ]
      summary: Create an action in a step
      operationId: createStepAction
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepActionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/issuances/{issuanceFlowId}/steps/{stepId}/actions/{actionId}:
    get:
      tags: [ "Issuance Flow" ]
      summary: Retrieve a single action
      operationId: getStepAction
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Issuance Flow" ]
      summary: Overwrite an action
      operationId: updateStepAction
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepActionRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Issuance Flow" ]
      summary: Delete an action
      operationId: deleteStepAction
      parameters:
        - name: issuanceFlowId
          in: path
          required: true
          schema:
            type: string
        - name: stepId
          in: path
          required: true
          schema:
            type: string
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: <https://auth.example.com/oauth/authorize>
          tokenUrl: <https://auth.example.com/oauth/token>
          scopes:
            editor: Grants editor access
            admin: Grants admin access
            viewer: Grants viewer access
    OIDC:
      type: openIdConnect
      openIdConnectUrl: <https://auth.example.com/.well-known/openid-configuration>
      description: OpenID Connect security scheme for OAuth2 flows
  schemas:
    Asset:
      type: object
      required:
        - id
        - mediaType
        - content
      properties:
        id:
          type: string
          description: Unique identifier for the asset in the database
          example: 123e4567-e89b-12d3-a456-426614174469
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
...
    Step:
      type: object
      required:
        - id
        - title
        - description
        - order
        - type
        - actions
      properties:
        id:
          type: string
          description: Unique identifier for the step
          example: 123e4567-e89b-12d3-a456-434314174000
        title:
          type: string
          description: Title of the step
          example: Verify Identity
        description:
          type: string
          description: Detailed description of the step
          example: Verify the user's identity
        order:
          type: integer
          description: Order of the step in the workflow
          minimum: 0
          example: 1
        type:
          $ref: '#/components/schemas/StepType'
        subFlow:
          type: string
          description: Optional sub-workflow for this step
          example: 123e4567-e89b-12d3-a456-434314174000
        actions:
          type: array
          items:
            $ref: '#/components/schemas/StepAction'
        asset:
          $ref: '#/components/schemas/Asset'
    StepRequest:
      type: object
      required:
        - title
        - description
        - order
        - type
        - actions
      properties:
        title:
          type: string
          description: Title of the step
          example: Verify Identity
        description:
          type: string
          description: Detailed description of the step
          example: Verify the user's identity
        order:
          type: integer
          description: Order of the step in the workflow
          minimum: 0
          example: 1
        type:
          $ref: '#/components/schemas/StepType'
        subFlow:
          type: string
          description: Optional sub-workflow for this step
          example: 123e4567-e89b-12d3-a456-434314174000
        actions:
          type: array
          description: List of actions associated with this step
          items:
            $ref: '#/components/schemas/StepActionRequest'
        asset:
          type: string
          description: Asset referenced by this step
          example: 456e4567-e89b-12d3-a456-426614174000
    StepsResponse:
      properties:
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
...
    StepType:
      type: string
      enum:
        - HUMAN_TASK
        - SERVICE
        - WORKFLOW
      description: Type of step in the workflow
      example: HUMAN_TASK
    StepAction:
      type: object
      required:
        - id
        - actionType
        - title
        - text
      properties:
        id:
          type: string
          description: Unique identifier for the action
          example: 123e4567-ef2d-12d3-abcd-426614174456
        actionType:
          type: string
          description: Type of action
          example: EXAMPLE_ACTION
        title:
          type: string
          description: Title of the action
          example: Connect Wallet
        text:
          type: string
          description: Descriptive text for the action
          example: Connect your wallet to continue
      discriminator:
        propertyName: actionType
        mapping:
          ARIES_OOB: '#/components/schemas/AriesOOBAction'
    StepActionRequest:
      type: object
      required:
        - actionType
        - title
        - text
      properties:
        title:
          type: string
          description: Title of the action
          example: Connect Wallet
        text:
          type: string
          description: Descriptive text for the action
          example: Connect your wallet to continue
        actionType:
          type: string
          example: HUMAN_TASK
      discriminator:
        propertyName: actionType
        mapping:
          ARIES_OOB: '#/components/schemas/AriesOOBAction'
    StepActionsResponse:
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/StepAction'
...
    AriesOOBAction:
      allOf:
        - $ref: '#/components/schemas/StepAction'
        - type: object
          required:
            - proofRequest
          properties:
            proofRequest:
              $ref: '#/components/schemas/AriesProofRequest'
    AriesProofRequest:
      type: object
      properties:
        attributes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AriesRequestCredentialAttributes'
          description: Map of attribute groups requested in the proof
          example:
            name:
              value: John Doe
        predicates:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AriesRequestCredentialPredicates'
          description: Map of predicate groups requested in the proof
          example:
            age:
              value: 30
    AriesRequestCredentialAttributes:
      type: object
      properties:
        attributes:
          type: array
          items:
            type: string
            description: Name of the attribute to be included in the proof
            example: name
          description: List of attribute names to be included in the proof
        restrictions:
          type: array
          items:
            type: string
            description: Name of the restriction to be included in the proof
            example: age
          description: List of restrictions on the requested attributes
...
    Persona:
      type: object
      required:
        - name
        - role
      properties:
        id:
          type: string
          description: Unique identifier for the persona
          example: 123e4567-e89b-12d3-a456-426614174456
        name:
          type: string
          description: Name of the persona
          example: John Doe
        role:
          type: string
          description: Role of the persona in the system
          example: Verifier
        description:
          type: string
          description: Detailed description of the persona
          example: John Doe is a verifier for the system
        headshotImage:
          $ref: '#/components/schemas/Asset'
        bodyImage:
          $ref: '#/components/schemas/Asset'
...
    Workflow:
      type: object
      required:
        - name
        - type
      properties:
        id:
          type: string
          description: Unique identifier for the workflow
          example: 789e4567-e89b-12d3-a456-434314174123
        name:
          type: string
          description: Name of the workflow
          example: Credential Issuance
        description:
          type: string
          description: Detailed description of the workflow
          example: This workflow issues credentials to users
        type:
          type: string
          enum: [ ISSUANCE, PRESENTATION ]
          description: Type of workflow
          example: ISSUANCE
        steps:
          type: array
          description: Ordered list of steps in the workflow
          items:
            $ref: '#/components/schemas/Step'
        personas:
          type: array
          description: References to personas used in this workflow
          items:
            $ref: '#/components/schemas/Persona'
    WorkflowRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Name of the workflow
          example: Credential Issuance
        description:
          type: string
          description: Detailed description of the workflow
          example: This workflow issues credentials to users
        steps:
          type: array
          description: Ordered list of steps in the workflow
          items:
            $ref: '#/components/schemas/StepRequest'
        personas:
          type: array
          description: References to personas used in this workflow
          items:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174456
    IssuanceFlowsResponse:
      type: object
      properties:
        issuanceFlows:
          type: array
          items:
            $ref: '#/components/schemas/IssuanceFlow'
    IssuanceFlowResponse:
      type: object
      properties:
        issuanceFlow:
          $ref: '#/components/schemas/IssuanceFlow'
    IssuanceFlow:
      type: object
      allOf:
        - $ref: '#/components/schemas/Workflow'
        - type: object
          required:
            - issuer
          properties:
            issuer:
              $ref: '#/components/schemas/Issuer'
      description: Workflow specialization for credential issuance
    IssuanceFlowRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/WorkflowRequest'
        - type: object
          required:
            - issuer
          properties:
            issuer:
              type: string
              example: "313e4567-e89b-12d3-a456-426614174469"
      description: Workflow specialization for credential issuance
...
    IdentifierType:
      type: string
      enum:
        - DID
      description: Types of supported identifiers
      example: DID
    IssuerType:
      type: string
      enum:
        - ARIES
      description: Types of supported issuers
...
    Issuer:
      type: object
      required:
        - id
        - name
        - description
        - type
        - credentialDefinitions
      properties:
        id:
          type: string
          description: Unique identifier for the issuer
          example: 123e4567-e89b-12d3-abcd-426614174456
        name:
          type: string
          description: Name of the issuer
          example: Issuer Name
        description:
          type: string
          description: Detailed description of the issuer
          example: This issuer issues credentials to users
        type:
          $ref: '#/components/schemas/IssuerType'
        identifierType:
          $ref: '#/components/schemas/IdentifierType'
        identifier:
          type: string
          description: External identifier of this issuer
          example: did:sov:XUeUZauFLeBNofY3NhaZCB
        organization:
          type: string
          description: Organization the issuer belongs to
          example: Acme Corporation
        logo:
          $ref: '#/components/schemas/Asset'
        credentialDefinitions:
          type: array
          items:
            $ref: '#/components/schemas/CredentialDefinition'
        credentialSchema:
          type: array
          items:
            $ref: '#/components/schemas/CredentialSchema'
...
    CredentialType:
      type: string
      enum:
        - ANONCRED
      description: Types of supported credentials
...
    CredentialDefinition:
      type: object
      required:
        - id
        - name
        - version
        - type
        - schemaId
        - icon
        - representations
      properties:
        id:
          type: string
          description: Unique identifier for the credential definition
          example: 123e4567-e89b-12d3-a456-426614174123
        name:
          type: string
          description: Name of the credential definition
          example: Credential Definition Name
        issuerId:
          type: string
          description: Unique identifier of the issuer who created this schema
          example: 123e4567-e89b-12d3-a456-426614174123
        schemaId:
          type: string
          description:  Identifier of the associated credential schema
          example:  123e4567-e89b-12d3-a456-426614174123
        version:
          type: string
          description: Version of the credential definition
          example: '1.0'
        type:
          $ref: '#/components/schemas/CredentialType'
        representations:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/CredentialRepresentation'
              - $ref: '#/components/schemas/OCARepresentation'
        revocation:
          oneOf:
            - $ref: '#/components/schemas/RevocationInfo'
            - $ref: '#/components/schemas/AnonCredRevocation'
        icon:
          $ref: '#/components/schemas/Asset'
...
    CredentialRepresentation:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier for the credential representation
          example: 123e4567-e89b-12d3-abcd-426614174456
...
  responses:
    'BadRequest':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid request data'
    'Unauthorized':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid authentication credentials'
    'Forbidden':
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Insufficient permissions to access resource'
    'NotFound':
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Resource not found'
    'InternalServerError':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'An unexpected error occurred'